added changes to structure for cleanliness
