--- src/instrument/configs/README.md
@@ -1,3 +1,0 @@
-# Configuration files
-
-- `iconfig.yml` - configuration for data collection
