+++ src/instrument/configs/__init__.py
@@ -1,0 +1,1 @@
+"""Configs required to set up user package"""
